+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ## [0.8] - unreleased
 ### Added
-- **examples/spectrum**: audio example
+- **examples/audio/**: example application for driving a PCA9685 via realtime audio
 
 ### Changed
 
@@ -12,7 +12,6 @@
 
 ## [0.7] - 2018-05-27
 ### Added
-- **examples/audio/**: example application for driving a PCA9685 via realtime audio
 - **examples/olaclient/**: example application for driving a PCA9685 via DMX512
 - **netinst.sh**: build script suitable to be run as an EXTRA from `picontrol-netinst`